requirements.txt
yt-dlp
ffmpeg-python
assemblyai
python-decouple
moviepy

